include duplicate NADs for UDN
normally the controller creates the NAD for a UDN
and that is sufficient. in this case we are
creating duplicates of the NAD so the duplicates
can be used to attach secondary NICs to the same segment

this is necessary because OVN-K8s does not support
two attachments to the same NAD. Linux Bridge does.